Migrating the native messaging protocol to a 64bit-wide version variable. Allows to increment, practically indefinitely, various builds/version numbers for the NetCPU processor-mechanisms thereby effectively filtering any of the accidentally-old/version-trailing processors connecting from the older comps/boxes... To preserve backwards compatibility (i.e. to allow older/stale versions of the messaging protocol to be updated automatically when such boxes go online), the port-number for the new update_server has been changed (with the older-compatible version still running on the old port, serving out the updated binaries to older-connecting boxes).

+++ netcpu/message.h
@@ -947,7 +947,7 @@ struct array<T, false> : field_interface {
 };
 
 struct version : message_base<message::messages_ids::version> {
-	message::scalar<uint8_t> value;
+	message::scalar<uint64_t> value;
 };
 
 struct ready_to_process : message_base<message::messages_ids::ready_to_process>  {
